source/requirements.txt
editdistance==0.5.3
imutils==0.5.4
matplotlib==3.3.4
numpy==1.19.2
opencv-python==3.4.10.35
pillow==8.1.2
pyspellchecker==0.6.2
pyqt5==5.15.2
qt-material==2.8.8
tensorflow==2.4.1
